FIAP/Cap 3 - (IR ALÉM) Implementando algoritmos de Machine Learning com Scikit-learn/requirements.txt
scikit-learn>=1.1
pandas>=1.3
numpy>=1.21
pyyaml
joblib
matplotlib
fastapi
uvicorn
shap
pytest
seaborn